+++ Editor/Common/ImagePacker/ImagePacker.cs
@@ -17,12 +17,13 @@ namespace UnityEditor.U2D.Aseprite.Common
         /// </summary>
         /// <param name="rects">Rects to pack</param>
         /// <param name="padding">Padding between each rect</param>
+        /// <param name="requireSquarePOT">If true, the final texture will be power of two with equally sized sides</param>
         /// <param name="outPackedRects">Rects arranged within outPackedWidth and outPackedHeight</param>
         /// <param name="outPackedWidth">Width of the packed rects</param>
         /// <param name="outPackedHeight">Height of the packed rects</param>
-        public static void Pack(RectInt[] rects, int padding, out RectInt[] outPackedRects, out int outPackedWidth, out int outPackedHeight)
+        public static void Pack(RectInt[] rects, int padding, bool requireSquarePOT, out RectInt[] outPackedRects, out int outPackedWidth, out int outPackedHeight)
         {
-            var packNode = InternalPack(rects, padding);
+            var packNode = InternalPack(rects, padding, requireSquarePOT);
             outPackedWidth = packNode.rect.width;
             outPackedHeight = packNode.rect.height;
             var visitor = new CollectPackNodePositionVisitor();
@@ -48,12 +49,13 @@ namespace UnityEditor.U2D.Aseprite.Common
         /// <param name="height">Image buffers height</param>
         /// <param name="padding">Padding between each packed image</param>
         /// <param name="spriteSizeExpand">Pack sprite expand size</param>
+        /// <param name="requireSquarePOT">If true, the final texture will be power of two with equally sized sides</param>
         /// <param name="outPackedBuffer">Packed image buffer</param>
         /// <param name="outPackedBufferWidth">Packed image buffer's width</param>
         /// <param name="outPackedBufferHeight">Packed image buffer's height</param>
         /// <param name="outPackedRect">Location of each image buffers in the packed buffer</param>
         /// <param name="outUVTransform">Translation data from image original buffer to packed buffer</param>
-        public static void Pack(NativeArray<Color32>[] buffers, int[] width, int[] height, int padding, uint spriteSizeExpand, out NativeArray<Color32> outPackedBuffer, out int outPackedBufferWidth, out int outPackedBufferHeight, out RectInt[] outPackedRect, out Vector2Int[] outUVTransform)
+        public static void Pack(NativeArray<Color32>[] buffers, int[] width, int[] height, int padding, uint spriteSizeExpand, bool requireSquarePOT, out NativeArray<Color32> outPackedBuffer, out int outPackedBufferWidth, out int outPackedBufferHeight, out RectInt[] outPackedRect, out Vector2Int[] outUVTransform)
         {
             UnityEngine.Profiling.Profiler.BeginSample("Pack");
             // Determine the area that contains data in the buffer
@@ -69,7 +71,7 @@ namespace UnityEditor.U2D.Aseprite.Common
                     t.height = tightRects[i].height + (int)spriteSizeExpand;
                     tightRectArea[i] = t;
                 }
-                Pack(tightRectArea, padding, out outPackedRect, out outPackedBufferWidth, out outPackedBufferHeight);
+                Pack(tightRectArea, padding, requireSquarePOT, out outPackedRect, out outPackedBufferWidth, out outPackedBufferHeight);
                 var packBufferSize = (ulong)outPackedBufferWidth * (ulong)outPackedBufferHeight;
 
                 if (packBufferSize < 0 || packBufferSize >= int.MaxValue)
@@ -97,23 +99,27 @@ namespace UnityEditor.U2D.Aseprite.Common
             }
         }
 
-        static ImagePackNode InternalPack(RectInt[] rects, int padding)
+        static ImagePackNode InternalPack(RectInt[] rects, int padding, bool requireSquarePOT)
         {
             if (rects == null || rects.Length == 0)
                 return new ImagePackNode() { rect = new RectInt(0, 0, 0, 0)};
             var sortedRects = new ImagePackRect[rects.Length];
-            for (int i = 0; i < rects.Length; ++i)
+            for (var i = 0; i < rects.Length; ++i)
             {
                 sortedRects[i] = new ImagePackRect();
                 sortedRects[i].rect = rects[i];
                 sortedRects[i].index = i;
             }
+            var initialWidth = (int)NextPowerOfTwo((ulong)rects[0].width);
             var initialHeight = (int)NextPowerOfTwo((ulong)rects[0].height);
+            if (requireSquarePOT)
+                initialWidth = initialHeight = (initialWidth > initialHeight) ? initialWidth : initialHeight;
+
             Array.Sort<ImagePackRect>(sortedRects);
             var root = new ImagePackNode();
-            root.rect = new RectInt(0, 0, (int)NextPowerOfTwo((ulong)rects[0].width), initialHeight);
+            root.rect = new RectInt(0, 0, initialWidth, initialHeight);
 
-            for (int i = 0; i < rects.Length; ++i)
+            for (var i = 0; i < rects.Length; ++i)
             {
                 if (!root.Insert(sortedRects[i], padding)) // we can't fit
                 {
@@ -126,6 +132,10 @@ namespace UnityEditor.U2D.Aseprite.Common
                     }
                     else
                         newHeight = (int)NextPowerOfTwo((ulong)root.rect.height + 1);
+
+                    if (requireSquarePOT)
+                        newWidth = newHeight = (newWidth > newHeight) ? newWidth : newHeight;
+
                     // Reset all packing and try again
                     root = new ImagePackNode();
                     root.rect = new RectInt(0, 0, newWidth, newHeight);